there is now a PoserClientBase, poser clients are like 2 lines now
