+++ virtualreality/templates/__init__.py
@@ -2,3 +2,4 @@
 from .poses import *
 from .templates import *
 from .udu_templates import *
+from .template_base import *
