+++ virtualreality/templates/template_base.py
@@ -71,7 +71,7 @@ class PoserTemplateBase(object):
 
     async def send(self):
         """Send all poses thread."""
-        raise NotImplementedError('please implement the send thread to send pose packets')
+        raise NotImplementedError('please implement the send thread')
 
     async def recv(self):
         """Receive messages thread."""
@@ -174,3 +174,54 @@ class PoserTemplateBase(object):
             return _thread_reg_wrapper
 
         return _thread_reg
+
+class PoserClientBase(PoserTemplateBase):
+    '''
+    poser client base class
+    should only be inherited together with one of the poser templates
+    should only be used to crate new poser clients
+    '''
+    def __init__(self, *args, **kwargs):
+        """init"""
+        super().__init__(*args, **kwargs)
+        self._coro_name_exceptions.append("thread_register")
+
+    def thread_register(self, sleep_delay, runInDefaultExecutor=False):
+        """
+        Registers a thread for the client
+
+        sleepDelay - sleep delay in seconds
+        runInDefaultExecutor - bool, set True if you want the function to be executed in asyncio's default pool executor
+        """
+
+        def _thread_register(coro):
+            if not asyncio.iscoroutinefunction(coro) and not runInDefaultExecutor:
+                raise ValueError(f"{repr(coro)} is not a coroutine function and runInDefaultExecutor is set to False")
+
+            if coro.__name__ not in self.coro_keep_alive and coro.__name__ not in self.coro_list:
+                self.coro_keep_alive[coro.__name__] = KeepAliveTrigger(True, sleep_delay)
+                self.coro_list.append(coro.__name__)
+
+                if runInDefaultExecutor:
+
+                    def _wrapper(*args, **kwargs):
+                        setattr(_wrapper, "__name__", f"{coro.__name__} _decorated")
+                        loop = asyncio.get_running_loop()
+
+                        return loop.run_in_executor(None, coro, *args, **kwargs)
+
+                    setattr(_wrapper, "__name__", coro.__name__)
+                    setattr(self, coro.__name__, _wrapper)
+                    return _wrapper
+
+                else:
+                    setattr(self, coro.__name__, coro)
+
+            else:
+                raise NameError(
+                    f"trying to register existing thread, thread with name {repr(coro.__name__)} already exists"
+                )
+
+            return coro
+
+        return _thread_register
