+++ virtualreality/templates/templates.py
@@ -42,10 +42,10 @@ class PoserTemplate(PoserTemplateBase):
 
             @PoserTemplate.register_member_thread(0.5)
             async def example_thread(self):
-                while self.coro_keep_alive['example_thread'][0]:
+                while self.coro_keep_alive['example_thread'].is_alive:
                     self.pose.x += 0.04
 
-                    await asyncio.sleep(self.coro_keep_alive['example_thread'][1])
+                    await asyncio.sleep(self.coro_keep_alive['example_thread'].sleep_delay)
 
         poser = MyPoser()
 
@@ -95,7 +95,7 @@ class PoserTemplate(PoserTemplateBase):
                 break
 
 
-class PoserClient(PoserTemplate):
+class PoserClient(PoserTemplate, PoserClientBase):
     """
     PoserClient.
 
@@ -104,55 +104,15 @@ class PoserClient(PoserTemplate):
 
         @poser.thread_register(1)
         async def lol():
-            while poser.coro_keep_alive['lol'][0]:
+            while poser.coro_keep_alive['lol'].is_alive:
                 poser.pose.x += 0.2
 
-                await asyncio.sleep(poser.coro_keep_alive['lol'][1])
+                await asyncio.sleep(poser.coro_keep_alive['lol'].sleep_delay)
 
         asyncio.run(poser.main())
     """
 
     def __init__(self, *args, **kwargs):
         """init"""
+        print ('starting the client')
         super().__init__(*args, **kwargs)
-        self._coro_name_exceptions.append("thread_register")
-
-    def thread_register(self, sleep_delay, runInDefaultExecutor=False):
-        """
-        Register a thread for PoserClient.
-
-        sleepDelay - sleep delay in seconds
-        runInDefaultExecutor - bool, set True if you want the function to be executed in asyncio's default pool executor
-        """
-
-        def _thread_register(coro):
-            if not asyncio.iscoroutinefunction(coro) and not runInDefaultExecutor:
-                raise ValueError(f"{repr(coro)} is not a coroutine function and runInDefaultExecutor is set to False")
-
-            if coro.__name__ not in self.coro_keep_alive and coro.__name__ not in self.coro_list:
-                self.coro_keep_alive[coro.__name__] = KeepAliveTrigger(True, sleep_delay)
-                self.coro_list.append(coro.__name__)
-
-                if runInDefaultExecutor:
-
-                    def _wrapper(*args, **kwargs):
-                        setattr(_wrapper, "__name__", f"{coro.__name__} _decorated")
-                        loop = asyncio.get_running_loop()
-
-                        return loop.run_in_executor(None, coro, *args, **kwargs)
-
-                    setattr(_wrapper, "__name__", coro.__name__)
-                    setattr(self, coro.__name__, _wrapper)
-                    return _wrapper
-
-                else:
-                    setattr(self, coro.__name__, coro)
-
-            else:
-                raise NameError(
-                    f"trying to register existing thread, thread with name {repr(coro.__name__)} already exists"
-                )
-
-            return coro
-
-        return _thread_register
